requirements.txt
Django>=5,<6
djangorestframework>=3.15
celery>=5.4
redis>=5.0
feedparser>=6.0
python-dotenv>=1.0
